feat: Implement bilingual support for settings page
- Add comprehensive translation keys for settings page content
- Create SettingsPageContent component with proper bilingual support
- Replace all hardcoded text with translation references
- Update English and Spanish translation files with settings-specific keys
- Add TypeScript type definitions for new translation keys
- Maintain mobile-first UX design and 44px touch targets
- Ensure proper fallbacks for all translated content

Translation keys added:
- settings.title, settings.subtitle
- settings.profileDescription
- settings.noCredentials, settings.noCredentialsDescription
- settings.failedToLoadCredentials, settings.failedToRemoveCredential
- common.na for "N/A" in both languages
