+++ src/app/(dashboard)/dashboard/components/index.ts
@@ -1,0 +1,6 @@
+export { DashboardHeader } from './DashboardHeader'
+export { DashboardMetrics } from './DashboardMetrics'
+export { CriticalActionsAlert } from './CriticalActionsAlert'
+export { OverdueDraftsCard } from './OverdueDraftsCard'
+export { PendingApprovalsCard } from './PendingApprovalsCard'
+export { AdministrativeActionsPanel } from './AdministrativeActionsPanel'
