+++ src/app/(dashboard)/settings/SettingsClient.tsx
@@ -36,11 +36,11 @@ export default function SettingsClient({ user }: SettingsClientProps) {
       if (result.success) {
         setCredentials(result.credentials || [])
       } else {
-        showError(result.error || 'Failed to load credentials')
+        showError(result.error || t.settings?.failedToLoadCredentials || 'Failed to load credentials')
       }
     } catch (err) {
       console.error('Error loading credentials:', err)
-      showError('Failed to load credentials')
+      showError(t.settings?.failedToLoadCredentials || 'Failed to load credentials')
     } finally {
       setLoading(false)
     }
@@ -77,11 +77,11 @@ export default function SettingsClient({ user }: SettingsClientProps) {
         success(t.biometric?.removeSuccess || 'Biometric credential removed successfully')
         setCredentials(prev => prev.filter(cred => cred.credentialId !== credentialId))
       } else {
-        showError(result.error || 'Failed to remove credential')
+        showError(result.error || t.settings?.failedToRemoveCredential || 'Failed to remove credential')
       }
     } catch (err) {
       console.error('Error removing credential:', err)
-      showError('Failed to remove credential')
+      showError(t.settings?.failedToRemoveCredential || 'Failed to remove credential')
     } finally {
       setRemoving(null)
     }
@@ -132,56 +132,75 @@ export default function SettingsClient({ user }: SettingsClientProps) {
 
   return (
     <div className="space-y-6">
-      {/* User Profile Section */}
-      <div className="bg-white shadow rounded-lg">
-        <div className="px-4 py-3 border-b border-gray-200">
-          <div className="flex items-center justify-between">
-            <h2 className="text-lg font-semibold text-gray-900">
-              {t.settings?.profile || 'Profile'}
-            </h2>
+      {/* User Profile Section - Mobile-First Design */}
+      <div className="bg-white rounded-2xl border border-gray-200/60 shadow-sm">
+        <div className="px-6 py-5 border-b border-gray-200/50">
+          <div className="flex flex-col sm:flex-row sm:items-center sm:justify-between gap-4">
+            <div>
+              <h2 className="text-lg font-bold text-gray-900">
+                {t.settings?.profile || 'Profile'}
+              </h2>
+              <p className="text-sm text-gray-600 mt-1">{t.settings?.profileDescription || 'Your personal information and account details'}</p>
+            </div>
             <button
               onClick={() => setShowChangePassword(true)}
-              className="px-3 py-1.5 text-sm font-medium text-blue-600 bg-blue-50 border border-blue-200 rounded-lg hover:bg-blue-100 active:scale-95 transition-all duration-150 touch-manipulation"
+              className="inline-flex items-center justify-center px-4 py-3 min-h-[44px] text-sm font-semibold text-primary bg-primary/10 border border-primary/20 rounded-xl hover:bg-primary/20 hover:scale-105 active:scale-95 transition-all duration-200 touch-manipulation"
             >
+              <svg className="w-4 h-4 mr-2" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 15v2m-6 4h12a2 2 0 002-2v-6a2 2 0 00-2-2H6a2 2 0 00-2 2v6a2 2 0 002 2zm10-10V7a4 4 0 00-8 0v4h8z" />
+              </svg>
               {t.settings?.changePassword || 'Change Password'}
             </button>
           </div>
         </div>
-        <div className="px-4 py-3">
-          <div className="grid grid-cols-1 md:grid-cols-2 gap-3">
-            <div className="flex">
-              <div className="flex-1">
-                <label className="block text-xs font-bold text-gray-700">
+        <div className="px-6 py-6">
+          <div className="space-y-6">
+            {/* Name and Position Row */}
+            <div className="grid grid-cols-1 sm:grid-cols-2 gap-6">
+              <div>
+                <label className="block text-sm font-semibold text-gray-700 mb-2">
                   {t.auth?.name || 'Name'}
                 </label>
-                <p className="mt-0.5 text-sm text-gray-900">{user.name}</p>
+                <div className="p-3 bg-gray-50 rounded-xl border border-gray-200">
+                  <p className="text-sm text-gray-900 font-medium">{user.name}</p>
+                </div>
               </div>
-              <div className="flex-1">
-                <label className="block text-xs font-bold text-gray-700">
+              <div>
+                <label className="block text-sm font-semibold text-gray-700 mb-2">
                   {t.users?.position || 'Position'}
                 </label>
-                <p className="mt-0.5 text-sm text-gray-900">{user.position || 'N/A'}</p>
+                <div className="p-3 bg-gray-50 rounded-xl border border-gray-200">
+                  <p className="text-sm text-gray-900 font-medium">{user.position || t.common?.na || 'N/A'}</p>
+                </div>
               </div>
             </div>
+
+            {/* Email */}
             <div>
-              <label className="block text-xs font-bold text-gray-700">
+              <label className="block text-sm font-semibold text-gray-700 mb-2">
                 {t.auth?.email || 'Email'}
               </label>
-              <p className="mt-0.5 text-sm text-gray-900">{user.email}</p>
+              <div className="p-3 bg-gray-50 rounded-xl border border-gray-200">
+                <p className="text-sm text-gray-900 font-medium">{user.email}</p>
+              </div>
             </div>
-            <div className="md:col-span-2">
-              <div className="flex">
-                <div className="flex-1">
-                  <label className="block text-xs font-bold text-gray-700">
-                    {t.common?.department || 'Department'}
-                  </label>
-                  <p className="mt-0.5 text-sm text-gray-900">{user.department || 'N/A'}</p>
+
+            {/* Department and Role Row */}
+            <div className="grid grid-cols-1 sm:grid-cols-2 gap-6">
+              <div>
+                <label className="block text-sm font-semibold text-gray-700 mb-2">
+                  {t.common?.department || 'Department'}
+                </label>
+                <div className="p-3 bg-gray-50 rounded-xl border border-gray-200">
+                  <p className="text-sm text-gray-900 font-medium">{user.department || t.common?.na || 'N/A'}</p>
                 </div>
-                <div className="flex-1">
-                  <label className="block text-xs font-bold text-gray-700">
-                    {t.auth?.role || 'Role'}
-                  </label>
-                  <p className="mt-0.5 text-sm text-gray-900 capitalize">{user.role}</p>
+              </div>
+              <div>
+                <label className="block text-sm font-semibold text-gray-700 mb-2">
+                  {t.auth?.role || 'Role'}
+                </label>
+                <div className="p-3 bg-gray-50 rounded-xl border border-gray-200">
+                  <p className="text-sm text-gray-900 font-medium capitalize">{user.role}</p>
                 </div>
               </div>
             </div>
@@ -189,29 +208,40 @@ export default function SettingsClient({ user }: SettingsClientProps) {
         </div>
       </div>
 
-      {/* Biometric Authentication Section */}
-      <div className="bg-white shadow rounded-lg">
-        <div className="px-4 py-3 border-b border-gray-200">
-          <h2 className="text-lg font-semibold text-gray-900">
-            {t.biometric?.authentication || 'Biometric Authentication'}
-          </h2>
-          <p className="mt-1 text-xs text-gray-600">
-            {t.biometric?.settingsDescription || 'Manage your biometric authentication methods for secure login'}
-          </p>
+      {/* Biometric Authentication Section - Enhanced Mobile Design */}
+      <div className="bg-white rounded-2xl border border-gray-200/60 shadow-sm">
+        <div className="px-6 py-5 border-b border-gray-200/50">
+          <div className="flex items-start gap-4">
+            <div className="flex-shrink-0">
+              <div className="w-10 h-10 bg-green-100 rounded-xl flex items-center justify-center">
+                <svg className="w-5 h-5 text-green-600" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 15v2m-6 4h12a2 2 0 002-2v-6a2 2 0 00-2-2H6a2 2 0 00-2 2v6a2 2 0 002 2zm10-10V7a4 4 0 00-8 0v4h8z" />
+                </svg>
+              </div>
+            </div>
+            <div className="flex-1 min-w-0">
+              <h2 className="text-lg font-bold text-gray-900">
+                {t.biometric?.authentication || 'Biometric Authentication'}
+              </h2>
+              <p className="text-sm text-gray-600 mt-1">
+                {t.biometric?.settingsDescription || 'Manage your biometric authentication methods for secure login'}
+              </p>
+            </div>
+          </div>
         </div>
-        <div className="px-4 py-4">
+        <div className="px-6 py-6">
           {loading ? (
-            <div className="flex items-center justify-center p-4">
-              <LoadingSpinner size="sm" color="primary" />
+            <div className="flex items-center justify-center py-12">
+              <LoadingSpinner size="lg" color="primary" />
             </div>
           ) : (
-            <div className="space-y-4">
+            <div className="space-y-8">
               {/* Setup New Biometric */}
               <div>
-                <h3 className="text-base font-medium text-gray-900 mb-3">
+                <h3 className="text-base font-semibold text-gray-900 mb-4">
                   {t.biometric?.addNew || 'Add New Biometric Authentication'}
                 </h3>
-                <div className="flex justify-center">
+                <div className="bg-gray-50 rounded-2xl p-6 border border-gray-200">
                   <BiometricAuth
                     mode="setup"
                     userId={user.id}
@@ -219,7 +249,7 @@ export default function SettingsClient({ user }: SettingsClientProps) {
                     userDisplayName={user.name}
                     onSuccess={handleBiometricSetupSuccess}
                     onError={handleBiometricSetupError}
-                    className="max-w-sm"
+                    className="w-full max-w-none"
                   />
                 </div>
               </div>
@@ -227,38 +257,47 @@ export default function SettingsClient({ user }: SettingsClientProps) {
               {/* Existing Credentials */}
               {credentials.length > 0 && (
                 <div>
-                  <h3 className="text-base font-medium text-gray-900 mb-3">
+                  <h3 className="text-base font-semibold text-gray-900 mb-4">
                     {t.biometric?.existingCredentials || 'Your Biometric Credentials'}
                   </h3>
                   <div className="space-y-3">
                     {credentials.map((credential) => (
                       <div
                         key={credential.id}
-                        className="flex items-center justify-between p-3 border border-gray-200 rounded-lg"
+                        className="flex items-center justify-between p-4 bg-gray-50 border border-gray-200 rounded-xl hover:bg-gray-100 transition-colors"
                       >
-                        <div className="flex items-center space-x-3">
-                          {getDeviceIcon(credential.deviceType)}
-                          <div>
-                            <h4 className="text-sm font-medium text-gray-900">
+                        <div className="flex items-center space-x-4 flex-1 min-w-0">
+                          <div className="flex-shrink-0">
+                            {getDeviceIcon(credential.deviceType)}
+                          </div>
+                          <div className="flex-1 min-w-0">
+                            <h4 className="text-sm font-semibold text-gray-900 truncate">
                               {credential.deviceName || getDeviceTypeName(credential.deviceType)}
                             </h4>
-                            <p className="text-xs text-gray-500">
-                              {t.common?.created || 'Created'}: {formatDate(credential.createdAt)}
-                            </p>
-                            <p className="text-xs text-gray-500">
-                              {t.common?.lastUsed || 'Last used'}: {formatDate(credential.lastUsed)}
-                            </p>
+                            <div className="space-y-1 mt-1">
+                              <p className="text-xs text-gray-600">
+                                {t.common?.created || 'Created'}: {formatDate(credential.createdAt)}
+                              </p>
+                              <p className="text-xs text-gray-600">
+                                {t.common?.lastUsed || 'Last used'}: {formatDate(credential.lastUsed)}
+                              </p>
+                            </div>
                           </div>
                         </div>
                         <button
                           onClick={() => handleRemoveCredential(credential.credentialId)}
                           disabled={removing === credential.credentialId}
-                          className="px-3 py-1 text-sm text-red-600 hover:text-red-700 hover:bg-red-50 rounded-md transition-colors disabled:opacity-50"
+                          className="flex items-center justify-center px-4 py-2 min-h-[44px] text-sm font-semibold text-red-600 bg-red-50 border border-red-200 rounded-xl hover:bg-red-100 hover:scale-105 active:scale-95 transition-all duration-200 disabled:opacity-50 disabled:cursor-not-allowed touch-manipulation"
                         >
                           {removing === credential.credentialId ? (
                             <LoadingSpinner size="sm" color="red" />
                           ) : (
-                            t.common?.remove || 'Remove'
+                            <>
+                              <svg className="w-4 h-4 mr-2" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16" />
+                              </svg>
+                              {t.common?.remove || 'Remove'}
+                            </>
                           )}
                         </button>
                       </div>
@@ -268,14 +307,17 @@ export default function SettingsClient({ user }: SettingsClientProps) {
               )}
 
               {credentials.length === 0 && (
-                <div className="text-center py-6">
-                  <div className="text-gray-400 mb-2">
-                    <svg className="mx-auto h-8 w-8" fill="none" stroke="currentColor" viewBox="0 0 24 24">
-                      <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 15v2m-6 4h12a2 2 0 002-2v-6a2 2 0 00-2-2H6a2 2 0 00-2 2v6a2 2 0 002 2zm10-10V7a4 4 0 00-8 0v4h8z"/>
+                <div className="text-center py-12 bg-gray-50 rounded-2xl border-2 border-dashed border-gray-300">
+                  <div className="text-gray-400 mb-4">
+                    <svg className="mx-auto h-12 w-12" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                      <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={1.5} d="M12 15v2m-6 4h12a2 2 0 002-2v-6a2 2 0 00-2-2H6a2 2 0 00-2 2v6a2 2 0 002 2zm10-10V7a4 4 0 00-8 0v4h8z"/>
                     </svg>
                   </div>
-                  <p className="text-xs text-gray-500">
-                    {t.biometric?.noCredentials || 'No biometric credentials set up yet'}
+                  <p className="text-sm text-gray-600 font-medium">
+                    {t.settings?.noCredentials || 'No biometric credentials set up yet'}
+                  </p>
+                  <p className="text-xs text-gray-500 mt-1">
+                    {t.settings?.noCredentialsDescription || 'Set up biometric authentication above to enhance your account security'}
                   </p>
                 </div>
               )}
